requirements.txt
pyyaml
msal